Feature: Generate auth for Account.User
Run the Phoenix `mix.gen.auth` generator to create basic user
authentication.

This creates an `Account` context, with a `User` schema and `users`
database table.

The only changes have been a few tweaks to the tests to explicitly test
the expiration dates of user tokens instead of using a magic date of 2+
years ago. Specifically:

* Logins: 60 days
* Confirmation & Password resets: 7 days