Fix: Ecto database not reachable on CI
The Github Action was having configuration issues on the Github Action
(GA). This was traced down to using a `socket_dir` instead of a
username/password combination for authentication; the GA service for
Postgres does not support this.

This is solved by explicitly setting the `socket_dir` to `nil` in the
configuration block for the GA.

To make the configuration more portable, the  values for the GA Postgres
configuration are extracted to environment variables. These variables
are now set in the workflow yml file.

This workflow has been tested locally using the `nektos/act` tool.
A brief description of using this tool should be included in a future
README.md or CONTRIBUTING.md update.