Add boostrap to assets pipeline
This commit adds Bootstrap to the assets pipeline, in preparation of
moving the default styling to Bootstrap.

Since we are using the SCSS version of Bootstrap, `:sass_dart` needs to
be added to the `mix.exs` to compile the stylesheets.

A watcher is added to the development server, and sass compile stage is
added to the `assets.deploy` mix task.