Extract and style navbar for bootstrap
Extracted the navbar from the root template to a separate `_navbar`
partial.

Further extracted the main menu from the navbar to a
`_navbar_main_menu` partial. The user menu has been renamed
`_navbar_user_menu` to match.

Both menus and the parent navbar have been styled in a basic Bootstrap
style. This extraction and styling should make it easier for future
changes to extend the navbar.