Extract and style flash alerts
Extract the flash alerts to a new partial. This partial now checks if
there are any flashes; if so, a container div is displayed with flashes
of each type.

The container for an individual flash alert has also been extracted.
This partial displays a dismissible alert in the given style
`@alert_type`, and with the given content `@alert_content`.