Update pages/index.html
Update the template for pages/index.html, the main landing page for the
application. This page now implements a basic Bootstrap landing page,
with the default Phoenix framework copy removed and replaced with:

* A simple Hero for our application
  - Including register/login buttons
* A list of (currently non-existent) resources
* A list of planned features advertising the app

This page is meant as a placeholder, but it's better than the default
copy, and at least *works* now that Bootstrap is replacing the default
phoenix css styles.

The `layout/app.html` layout has been updated with some extra vertical
margins on the `main` element.