Add a .tool-versions
Add a .tool-versions to allow ASDF (or a user's version management tool
of choice) to manage Elixir & OTP versions in development.

Future versions of the CI workflow may go with a matrix strategy for
testing, but for now we're keeping things simple and *officially*
supporting only the latest versions of our tools.