Updates to README
Updated the README from the stock Phoenix to a proper description of the
project.

Added detailed **Contributing** section and **Planned Features** todo
list.

Added a CI status badge to the README header.

Changed the `name` in `elixir-ci.yml` to better reflect what test suite
we're running (`ExUnit`) in the status badge.