Install Wallaby in Test & CI
Install Wallaby for feature testing. Because we want to be test-driven,
it would be nice to have full outside-in feature tests for our
golden-path.

This requires some configuration changes to get running in the
development environment, along with the creation of a `FeatureCase` test
support module analogous to `ConnCase`.

Wallaby (as configured) requires chromium (or chrome) & a matching
version of chromedriver. In order to install `chromium-browser` on
versions of `ubuntu-19.10` and higher (including the last two LTS
releases), our workflow needs to either download the debs directly or
install from snaps. As `snap install` doesn't work on the Docker image,
out workflow download the debs and install them manually.