Add user sign_in to Wallaby tests
Since most (all?) of our features will require uses to be signed in, we need to
give our feature tests the ability to sign in users.

We start with using ExMachina to provide factory functions to create
`:user` and other data structures. Because our *User* type uses hashed
passwords, we add a `set_password/0` function that can be used as
follows, taking the password defined in the factory's attributes:

    user = build(:user, attributes) |> set_password |> insert

We define a `sign_in/1` function that takes a user (as above) and
automates the login process for us. Since this function actually visits
the login page, it should be placed early in the feature test flow, as
below:

    user = build(:user) |> set_password |> insert
    session
    |> sign_in(user)
    # |> <your flow & assertions here>

The `sign_in/1` function required changes to the SQL sandbox in our test
environment, which have also been made.