Extract sign_in to FeatureHelpers
Since this function will be used in many different features tests,
we should extract it to a FeatureHelpers module.

We can either import this module in every feature test that needs it, or
include it as part of our FeatureCase. Since signing in will be a nearly
universal part of our system, we're going with the latter strategy for
now.