Added LiveView index/show for /chat
Added the "/chat" route and LiveView. This route displays an index of
chat rooms in a column on the left side of the main application view,
and fills the right side of the application with (currently) the Room's
description.

This will eventually be used to display the actual Room (with it's
messages) on the right, allowing a user to easily switch rooms without
reloading the whole application.