Redirect signed-in users to /chat
Signed-in users should be redirect to /chat from the landing page.

Squelch noisy JavaScript log messages. They were providing too much
noise when doing behavior testing with Wallaby.

Temporarily disable the Ecto Sandbox manual mode. I'm not sure if it's
necessary and it was causing errors in some tests with the redirection
in place.