Update live flash alerts
Update the root layout for live views.
Update the extracted flash alerts for live views

This finishes the refactoring of flash alerts for live views, and
restyles the alerts and main application to be full-width. This gives it
more of an application look-and-feel, and ensures that the alerts are
more integrated into the application proper's UI.