Extract list and show components from ChatLive
Extract a `ListComponent` and `ShowComponent` from `ChatLive`. The html
template was getting a bit large and had too many responsibilities, so
this cleans it up nicely.

Extract setting the modal properties `@return_to` and `@modal_title`
from the template to the component. This cleans up the html code nicely
and provides a clearer division of responsibility between *form* (the
template) and *function* (the elixir component).