Add `has_many :messages` to `Chat.Room`
Add the 'has_many :messages` association to `Chat.Room`

Add API function `Chat.preload_room_messages/1` API function to assist
in preloading a `Room`'s messages. This function *also* fetches the
`User` records for each `Message` in the room, allowing for easy
rendering of the message log. The API call gets appended to a retrieval
function, such as

    Chat.get_room!(<id>) |> Chat.preload_room_messages

This API makes it easy to compose with other functions (and makes those
other functions easier to use in isolation).

Add a `Chat.get_room_by_name/1` function. This will be useful in the
future when replacing the route `/chat/:id` with `/chat/:title`. We also
should decide on consistency in calling this field `title` or `name`; I
think I'll go with `name`, which will come in a following commit on this
branch.

Add messages to the :show component. This simply renders the messages as
a series of cards, with the user information in the footer. There is a
lot of room for improvement, but this works for the moment.