Add `.projections.json`
Add a `.projections.json` to allow for a better IDE experience.
The projects simply make it easier for a compatible IDE (vim, VSCode,
etc) to associate a test and implementation file, and switch between
them.

It also relates live view files with the HTML component file, if
available.

This can be considered a rudimentary version, which can be fleshed out
as needs progress; but as the app gets more components it will become
easier to keep tests & implementations synced with nice editor
integrations like this.