Add missing @doc to `Chat`
The last commit added functions to the `Chat` context but neglected to
add `@doc` blocks. For completeness sake, those have been added here.

Removed a `@tag :focus` that had been left in one of the tests by
mistake. The CI doesn't make use of it so it doesn't affect the checks,
but we still shouldn't leave them hanging around.