Add PubSub subscription/handling
Subscribe users to room-specific PubSub services. This is the core of
the real-time chat system.

Add an overflow wrapper to the message list. This has the effect of both
scrolling the chat as new messages appear, *and* keeping the scroll area
to the bottom, always focussed on new messages.

TODO: I need to go back and add tests for these changes, but I wanted to
get this core functionality working because a) it's new to me, and b) it
helps to inform me, going forward, how to approach these kinds of
problems.