Tweak styling and sizing of rooms
Add the room title to the room header

Restyle the room description in the room header

Better calculate the max-height of the message wrapper, to prevent
scrolling the entire page on smaller device heights. This works by
subtracting the heights of the form, room header, and navbar, along with
margins, from a `100vh` initial max. I'm actively looking for a better
way to to this, but for now this is better than before.