Preload user after `create_message` succeeds
Since we will often need to use the user info, we preload the user
associated with a `create_message` if the insert succeeds; and pass on
the error otherwise.

Eventually these preloads will be refined so they only preload the
*useful* information, and don't pass on the *private* information such
as hashed passwords. For now, the `HarmonyWeb` side of the application
is taking care of sanitizing this, but we should do this in the
`Harmony` side of the boundary before a merge into `main`.