Use phx-update="append" for messages
Instead of force-reloading all the messages in a room each time a
message comes in, we now use `phx-update="append"` and just send the new
messages across the socket.

With all of the previous bug fixes added, this leads to *much* faster
(perceptually instantaneous) updates across multiple devices within the
chat rooms.