Exclude flaky Wallaby tests on CI
The Wallaby tests suddenly stopped working on CI, including previously
good commits with no changes to the CI configuration or code.

This might be due to an upstream change in the Docker image, or
something else. The tests *do* still work in a local CI run using
`nektos/act`, which adds further to the mystery; it may have to do with
GitHub suddenly limiting memory on our CI runs or something similar.
Until we can track down what's causing these issues, we'll just tag and
exclude these tests on CI runs.

Contributers will still be able to run these tests locally; and
maintainers will have to do so before accepting pull requests; but it's
the best we can do in the short-term.