Fix message-list max-height too large
Fix the calculation of the `max-height` property on the message list to
account for all the other vertical elements.

So that the application screen should never have to scroll
vertically---only the message list---we want the max-height of the
message-list to be just *barely* large enough to take up all the
remaining vertical space after the navbar, room header, and text input
(and their paddings/margins). The previous calculation missed some of
the margins in the calculations.

There still might be some scrolling if the room header spans multiple
lines; this will be addressed in a future feature/fix branch.