Update sign_up test text field query
Update the `Query` in the sign_up feature test from css id selector to
`text_field`.

The id was being used because the text_field query matches on partial
text matches; thus the fields labelled "Password" and "Password
confirmation" both matched the first "Password" query. This was solved
by passing the `count: 2, at: 0` options, to specify that we expect a
second field to match as well, and want the (0-indexed) first.