bugfix: room_header when no rooms exist
The user registration tests brought to light several bugs when no rooms
exist

a) The `page_title` assigns could not be set to
`Chat.default_room().name` because `Chat.default_room()` was nil
b) The `<.room_header>` could not be rendered when `room` was nil

The first problem is solved by wrapping the call to `Chat.default_room()
in a case statement, and setting (or not setting) the `page_title`
assign as needed.

The second problem is solved by rendering the `<.room_header>`
conditional on a room existing. A future commit could add a hero section
or similar here to encourage a user to create the first room.