fix: redirect to /users/log_in on log_out
Because the root "/" page is authenticated now, if we try to redirect
there after a logout we get a spurious flash error message about the
page being unauthorized before being redirected to the login page.

A slightly better UI would be to redirect directly to the login page
after a logout. This avoids the error message and otherwise has the same
results.