Added dynamic timestamps to messages
Added a dynamic timestamp to messages, in a "time ago" format. The
initial text in absolute format before the JS hook takes over, allowing
users without Javascript and archival bots to have a more accurate
experience that doesn't get stale over time (e.g. if a user stays in a
room for a long time, the timestamps on new messages don't perpetually
say "just now").

Implementation is through a JS hook, with a hand-rolled timeAgo
function. I'm contemplating using an npm package for a better
implementation of the algorithm, particularly for locale support.