Chat.delete_message_by_id/2 broadcasts {:delete_message
The `Chat.delete_message_by_id/2` function now broadcasts
`{:delete_message, %Chat.Message{}}` to a PubSub topic for the room,
`chat_room:#{room.name}`.

The `stream_delete` calls have been moved from the `handle_event/3` to
a `handle_info/2` callback. This avoids attempting to remove a message
from the stream twice for the current user, once as the event handler
and once in the message handler.