Add role to users / Only admins can create rooms
Added the `:role` field to `Accounts.User` as an `Ecto.Enum`
Three values are currently supported: `[:admin, :member, :moderator]`.
The underlying Postgres type is a custom ENUM to enforce the values at
the database level in addition to the application layer.

The `Chat.create_room/2` function now explicitly pattern matches to
allow only the `:admin` role to create a room. Other roles will receive
an error tuple. The modal and button are only rendered for admins.