Refactor ChatRoomLive.Edit as a modal
Refactor `ChatRoomLive.Edit` as a modal in root `ChatRoomLive`
Doing this requires the changesets to both use an `as:` option so they
can be distinguished from each other. Otherwise, the `<.simple_form>`
component creates duplicate ids.

The backdrop for the modals is now transparent with a subtle blur, to
better show that it is a modal and not just another page

The delete room now uses a `data-confirm` A future feature will be a
more full-functioned confirmation modal (such as one requiring we type
the room's full name), with some big scary disclaimer text.

The now unneeded route and old LiveViews have been removed.