Add SocketHelpers to refactor socket tuple returns
Many Phoenix actions and event handlers pass a socket through a
complicated pipeline, then return the socket wrapped in a tuple. Code of
the following form is common

```
socket =
  socket
  |> assign(something: something, something_else: something_else)
  |> do_work()
  |> do_more_work

{:noreply, socket}
```

Follwing a tip by German Velasco, we've added some helpers to
`HarmonyWeb.SocketHelpers` and included them in all LiveViews through
the macros in `HarmonyWeb`. The `noreply(socket)` helpers takes a
socket and wraps it as `{:noreply, socket}` , and the `ok(socket)`
helper does the same with `:ok`. These helpers can be added to the end
of a pipeline, cleaning up the former code like this:

```
socket
|> assign(something: something, something_else: something_else)
|> do_work()
|> do_more_work
|> noreply()
```

It's a simple refactor that should aid readability of the code.