Move room forms to LiveComponents
The `ChattRoomLive` LiveView was getting a bit heavy, and it'll get
heavier if more functionality is moved to modals and the this "root"
liveview has the handle *all* events from those forms.

So instead, we're starting to offload some concerns to LiveComponents
that can handle the behaviors themselves. The two implemented here are
`RoomNewComponent` and `RoomEditComponent`, which now handle the event
logic for the room editing and creation forms.

This will allow the root `ChatRoomLive` to focus on handling
chat-specific events, slimming down the module.

I'm not 100% certain on the naming yet, I'll consider more options in
time.