Unread messages marker
Implemented an unread messages marker. The marker uses the metadata
stored in the user's RoomMembership to remember the last read message id
number, and makes use of the fact that message_id's are UUIDv7 --- that
is, they are monotonically increasing with microsecond resolution.

An `:unread_marker` atom is spliced into the list of messages from a
room, based on the current user's RoomMembership. No such marker is
spliced if the user is just a visitor (has no membership), if the
last_read id is the last message in the room, or if the last_read id is
`nil` (the user has just joined the room).

The last_read metadata is updated on loading a room and every time a new
message is received.