Sidebar has unread counts
The sidebar shows a count of unread messages in each joined room. It
increments in real time as new messages are sent to that room

There is currently a slight bug if unread messages are deleted before
the room is read; the count won't *decrement*, so it may not be entirely
accurate if there are a lot of messages being deleted.