Bug fixes for unread counts
There were a lot of subtleties with the unread counts handling
joining/leaving new rooms. Mostly they boiled down to this: when
joining/leaving a new room:

a.  We need to handle the PubSub subscriptions properly *outside* of
    `mount/3`, or else the user won't see new messages in the room *or*
    see an increase in the unread counts (until a refresh)
b.  We need to differentiate in `list_joined_rooms_with_unread_counts/1`
    having 0 unread because the user has *never* read the room, and 0
    unread because the user is up to date on the room
c.  Using the former, we only increment the unread count on rooms that
    the user has viewed at least once
d.  And of course, once the user has entered a room, swap the internal
    representation of this boolean

All of these issues *could* be solved by just refreshing after a user
joins a room, of course, but this results in a much better user experience.