Add migration/schema for user_profiles
Added a migration and schema for user_profiles.

The profiles `belong_to` user, and each user `has_one` profile.
Because of this, we want the migration to generate the profiles for any
existing users as well. We do this with the provided `execute`
statement, which for each user in the db:

* generates a uuid for the profile
* copies the user.id to the profile's user_id
* copies the username as the initial display_name
* copies the inserted_at timestamp from the user
* sets the updated_at timestamp to the migration time

The timestamps are chosen here to reflect that in the future, all
profiles should be created alongside their users (and thus should have
matching inserted_at timestamps), while the updated_at timestamps will
diverge naturally.