Create profiles with user registration
We'll need *every* user to have a profile, so creating a profile with
registration is the next step to implementing them.

We default the `display_name` to the `username` the user chose when
registering; but unlike the username, there will be more freedom with
the display name; and it won't be used in authentication.