Install DaisyUI, update CoreComponents.input/.button
Installed DaisyUI as an npm asset.

Manually set the `data-theme` for the root index to `light`. Until we
update more of the app to Daisy, it might otherwise lead to some
components having light/dark variations and others not, resulting in a
mess.

Updated the CoreComponents.input and CoreComponents.button to use
DaisyUI styling. These are heavily used so getting them working was the
first priority.