Update Tailwinds from 3.4.3 to 4.0.9
Some features of DaisyUI (such as themes) require the newer major
version of Tailwinds, so we are upgrading it.

We are enabling the 'light', 'dark', and 'cupcake' themes for now,
mostly for testing reasons.

The `light` theme is set as both the `default` and the `prefersdark`
until we are certain that all the visual glitches in the dark theme are
ironed out: otherwise the user OS might call for the dark theme and
receive a degraded UX.