Show message replies in the reply drawer
Added the functionality to show replies in the reply drawer.
Of course, we don't yet have any way to actually *send* the replies.

The implementation uses the `get_message_with_replies` to get message,
it's replies, and the users/profiles in a single `Repo.get()`. This
results in a constant 6 queries:

* One for the `%Message{}`
* One for all of the `%Reply{}`s
* One for the `%Message{}` author
* One for the `%Message{}` author's profile
* One for all of the `%Reply{}` authors
* One for all of the `%Reply{}` authors' profiles

Fixed a long-standing N+1 query that was hitting the database once for
each message in a room to get the user profiles. We now preload them
with the message. This required updating some tests that manually create
messages and then send a message to the PubSub, because that broadcast
needs the profile preloaded as well.