Users can create and delete replies
Added the core context functions for CRUD acts on `%Chat.Reply{}`,
including:

- `Chat.list_replies/1`: list replies given the parent `message_id`
- `Chat.change_reply/2`: return a changeset for a `%Chat.Reply{}`
- `Chat.create_reply/3`: creates a reply for the given
  `%Accounts.User{}` and `%Chat.Message{}`. Currently has no
  authorization logic, but this can be added when needed.
- `delete_reply/2`: deletes the reply, as a specific `%Accounts.User{}`.
  Checks if the user owns the reply before deleting it.`

Like with messages, the `create_reply` and `delete_reply` functions
broadcast a message on the PubSub topic for the **room** the reply is
in, of the form `{:new/delete_reply, message_id, reply}`. We send it to
the room because we may need to update the message items in the room
itself to show new replies as they come in. The `ChatRoomLive` view is
responsible for sending a `send_update/2` call to the replies live
component, which then handles its own `:replies` stream.

The `RepliesComponent` live component handles updates with the
`%{new_reply: xyz, message_id: id}` and `%{deleted_reply: xyz,
message_id: id}` assigns specially before the more general clauses, so
that these messages can be handled properly. The general clauses handle
the setup when the reply drawer is opened on a new message, along with
the initial empty state.