Show reply avatars in main room
We want a list of avatars of replying users beneath messages with
replies, along with a count of replies.

We start by preloading the replies on `Chat.list_messages/1`. We then
define the `reply_avatar_group` function component in
`MessageComponents`.

Since this will now update when replies are
added/removed, we add a `stream_insert(:messages, message)` to the event
handler for new/deleted replies.